#1381 added some casts for passing data into %lld's refactored neowiz's errno debugging fix added -Werror to BASECPPFLAGS refactored base64decode so -1's aren't assigned to unsigned chars

+++ include/rudiments/private/linkedlistinlines.h
@@ -239,7 +239,7 @@ void LINKEDLIST_CLASS::print() const {
 	uint64_t	i=0;
 	for (linkedlistnodetype *current=first; current;
 		current=(linkedlistnodetype *)current->getNext()) {
-		printf("index %lld: ",i);
+		printf("index %lld: ",(long long)i);
 		current->print();
 		printf("\n");
 		i++;
